Makes id types explicitly unsigned 64bit

+++ src/common.h
@@ -49,7 +49,7 @@ typedef enum {
 } variable_domain_type_t;
 
 typedef uint32_t variable_value_t;
-typedef size_t variable_id_t;
+typedef uint64_t variable_id_t;
 
 // enumeration for factor function types
 typedef enum FACTOR_FUCNTION_TYPE {
@@ -67,9 +67,9 @@ typedef enum FACTOR_FUCNTION_TYPE {
   FUNC_UNDEFINED = -1,
 } factor_function_type_t;
 
-typedef size_t factor_id_t;
+typedef uint64_t factor_id_t;
 
-typedef size_t weight_id_t;
+typedef uint64_t weight_id_t;
 typedef double weight_value_t;
 
 /**
